Alert box inputs added.

+++ app/src/main/java/my/bop/finalassignment/AdapterClinician.java
@@ -1,6 +1,7 @@
 package my.bop.finalassignment;
 
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -8,6 +9,7 @@ import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
 
+import androidx.appcompat.app.AlertDialog;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.finalassignment.R;
@@ -72,7 +74,27 @@ public class AdapterClinician extends RecyclerView.Adapter<RecyclerView.ViewHold
             deletebutton = (Button) itemView.findViewById(R.id.button3);
             deletebutton.setOnClickListener(new View.OnClickListener() {
                 public void onClick(View v) {
-                    deleteClinician();
+                    AlertDialog myQuittingDialogBox = new AlertDialog.Builder(v.getContext())
+                            // set message, title, and icon
+                            .setTitle("Delete")
+                            .setMessage("Do you want to Delete")
+                            .setIcon(R.drawable.ic_baseline_delete_forever_24)
+
+                            .setPositiveButton("Delete", (dialog, whichButton) -> {
+                                deleteClinician();
+                                dialog.dismiss();
+                            })
+                            .setNegativeButton("cancel", new DialogInterface.OnClickListener() {
+                                public void onClick(DialogInterface dialog, int which) {
+
+                                    dialog.dismiss();
+
+                                }
+                            })
+                            .create();
+
+                    myQuittingDialogBox.show();
+
                 }
             });
