+++ app/src/main/java/my/bop/finalassignment/AdapterNotes.java
@@ -1,6 +1,7 @@
 package my.bop.finalassignment;
 
 import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -9,6 +10,7 @@ import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import androidx.appcompat.app.AlertDialog;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.finalassignment.R;
@@ -78,7 +80,27 @@ public class AdapterNotes extends RecyclerView.Adapter<RecyclerView.ViewHolder>{
 
             button_des.setOnClickListener(new View.OnClickListener() {
                 public void onClick(View v) {
-                    deleteNote();
+
+                    AlertDialog myQuittingDialogBox = new AlertDialog.Builder(v.getContext())
+                            // set message, title, and icon
+                            .setTitle("Delete")
+                            .setMessage("Do you want to Delete")
+                            .setIcon(R.drawable.ic_baseline_delete_forever_24)
+
+                            .setPositiveButton("Delete", (dialog, whichButton) -> {
+                                deleteNote();
+                                dialog.dismiss();
+                            })
+                            .setNegativeButton("cancel", new DialogInterface.OnClickListener() {
+                                public void onClick(DialogInterface dialog, int which) {
+
+                                    dialog.dismiss();
+
+                                }
+                            })
+                            .create();
+
+                    myQuittingDialogBox.show();
                 }
             });
