+++ app/src/main/java/my/bop/finalassignment/AdminAdmissionFinalisePatientPage.java
@@ -2,11 +2,13 @@ package my.bop.finalassignment;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 import androidx.core.view.GravityCompat;
 import androidx.drawerlayout.widget.DrawerLayout;
 
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
@@ -442,13 +444,30 @@ public class AdminAdmissionFinalisePatientPage extends AppCompatActivity impleme
 
     public void deleteFeedBack(View view)
     {
-        deletefeedback();
-        Intent intent = new Intent(getApplicationContext(), AdminAdmissionFinalisePatientPage.class);
-        intent.putExtra("mrn", mrn);
-        intent.putExtra("admissionid", admissionid);
-        startActivity(intent);
+        AlertDialog myQuittingDialogBox = new AlertDialog.Builder(this)
+                // set message, title, and icon
+                .setTitle("Delete")
+                .setMessage("Do you want to Delete")
+                .setIcon(R.drawable.ic_baseline_delete_forever_24)
 
+                .setPositiveButton("Delete", (dialog, whichButton) -> {
+                    deletefeedback();
+                    Intent intent = new Intent(getApplicationContext(), AdminAdmissionFinalisePatientPage.class);
+                    intent.putExtra("mrn", mrn);
+                    intent.putExtra("admissionid", admissionid);
+                    startActivity(intent);
+                    dialog.dismiss();
+                })
+                .setNegativeButton("cancel", new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int which) {
 
+                        dialog.dismiss();
+
+                    }
+                })
+                .create();
+
+        myQuittingDialogBox.show();
     }
 
     public boolean deletefeedback()
@@ -488,18 +507,38 @@ public class AdminAdmissionFinalisePatientPage extends AppCompatActivity impleme
 
     public void deleteAdmission(View view)
     {
-        if(deleteAdmissionDatabase())
-        {
-            Intent intent = new Intent(getApplicationContext(), AdminAdmissionPage.class);
-            startActivity(intent);
-        }
-        else
-        {
-            Intent intent = new Intent(getApplicationContext(), AdminAdmissionFinalisePatientPage.class);
-            intent.putExtra("mrn", mrn);
-            intent.putExtra("admissionid", admissionid);
-            startActivity(intent);
-        }
+
+        AlertDialog myQuittingDialogBox = new AlertDialog.Builder(this)
+                // set message, title, and icon
+                .setTitle("Delete")
+                .setMessage("Do you want to Delete")
+                .setIcon(R.drawable.ic_baseline_delete_forever_24)
+
+                .setPositiveButton("Delete", (dialog, whichButton) -> {
+                    if(deleteAdmissionDatabase())
+                    {
+                        Intent intent = new Intent(getApplicationContext(), AdminAdmissionPage.class);
+                        startActivity(intent);
+                    }
+                    else
+                    {
+                        Intent intent = new Intent(getApplicationContext(), AdminAdmissionFinalisePatientPage.class);
+                        intent.putExtra("mrn", mrn);
+                        intent.putExtra("admissionid", admissionid);
+                        startActivity(intent);
+                    }
+                    dialog.dismiss();
+                })
+                .setNegativeButton("cancel", new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int which) {
+
+                        dialog.dismiss();
+
+                    }
+                })
+                .create();
+
+        myQuittingDialogBox.show();
     }
 
     private boolean deleteAdmissionDatabase()
