+++ app/src/main/java/my/bop/finalassignment/AdminAdmissionModifyPatientNotesPage.java
@@ -2,6 +2,7 @@ package my.bop.finalassignment;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 import androidx.core.view.GravityCompat;
@@ -9,6 +10,7 @@ import androidx.drawerlayout.widget.DrawerLayout;
 
 import android.app.DatePickerDialog;
 import android.app.TimePickerDialog;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
@@ -43,9 +45,6 @@ public class AdminAdmissionModifyPatientNotesPage extends AppCompatActivity impl
     String admissionid;
     String username;
     String role;
-    String bednumber2;
-    String concerned;
-    String gender;
     EditText date_time_in;
     String datetime3;
     DrawerLayout drawerLayout;
@@ -308,12 +307,31 @@ public class AdminAdmissionModifyPatientNotesPage extends AppCompatActivity impl
 
     public void DeleteNotes(View view)
     {
-        deletePatientInformation();
-        Intent intent = new Intent(getApplicationContext(), AdminAdmissionPatientNotesPage.class);
-        intent.putExtra("mrn", mrn);
-        intent.putExtra("admissionid", admissionid);
-        Toast.makeText(AdminAdmissionModifyPatientNotesPage.this, "Successful.", Toast.LENGTH_SHORT).show();
-        startActivity(intent);
+        AlertDialog myQuittingDialogBox = new AlertDialog.Builder(this)
+                // set message, title, and icon
+                .setTitle("Delete")
+                .setMessage("Do you want to Delete")
+                .setIcon(R.drawable.ic_baseline_delete_forever_24)
+
+                .setPositiveButton("Delete", (dialog, whichButton) -> {
+                    deletePatientInformation();
+                    Intent intent = new Intent(getApplicationContext(), AdminAdmissionPatientNotesPage.class);
+                    intent.putExtra("mrn", mrn);
+                    intent.putExtra("admissionid", admissionid);
+                    Toast.makeText(AdminAdmissionModifyPatientNotesPage.this, "Successful.", Toast.LENGTH_SHORT).show();
+                    startActivity(intent);
+                    dialog.dismiss();
+                })
+                .setNegativeButton("cancel", new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int which) {
+
+                        dialog.dismiss();
+
+                    }
+                })
+                .create();
+
+        myQuittingDialogBox.show();
     }
 
     private void showDateTimeDialog(final EditText date_time_in)
