+++ app/src/main/java/my/bop/finalassignment/AdminMedicationModifyPage.java
@@ -2,11 +2,13 @@ package my.bop.finalassignment;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 import androidx.core.view.GravityCompat;
 import androidx.drawerlayout.widget.DrawerLayout;
 
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
@@ -255,10 +257,30 @@ public class AdminMedicationModifyPage extends AppCompatActivity implements Navi
 
     public void DeleteMedicationButton(View view)
     {
-        if(deleteMedDatabase())
-        {
-            startActivity(new Intent(AdminMedicationModifyPage.this, AdminMedicationPage.class));
-        }
+        AlertDialog myQuittingDialogBox = new AlertDialog.Builder(this)
+                // set message, title, and icon
+                .setTitle("Delete")
+                .setMessage("Do you want to Delete")
+                .setIcon(R.drawable.ic_baseline_delete_forever_24)
+
+                .setPositiveButton("Delete", (dialog, whichButton) -> {
+                    if(deleteMedDatabase())
+                    {
+                        startActivity(new Intent(AdminMedicationModifyPage.this, AdminMedicationPage.class));
+                    }
+                    dialog.dismiss();
+                })
+                .setNegativeButton("cancel", new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int which) {
+
+                        dialog.dismiss();
+
+                    }
+                })
+                .create();
+
+        myQuittingDialogBox.show();
+
     }
 
     private boolean deleteMedDatabase()
@@ -296,10 +318,29 @@ public class AdminMedicationModifyPage extends AppCompatActivity implements Navi
     
     public void DeleteAllMedicationButton(View view)
     {
-        if(deleteMedsFromAll())
-        {
-            startActivity(new Intent(AdminMedicationModifyPage.this, AdminMedicationPage.class));
-        }
+        AlertDialog myQuittingDialogBox = new AlertDialog.Builder(this)
+                // set message, title, and icon
+                .setTitle("Delete")
+                .setMessage("Do you want to Delete")
+                .setIcon(R.drawable.ic_baseline_delete_forever_24)
+
+                .setPositiveButton("Delete", (dialog, whichButton) -> {
+                    if(deleteMedsFromAll())
+                    {
+                        startActivity(new Intent(AdminMedicationModifyPage.this, AdminMedicationPage.class));
+                    }
+                    dialog.dismiss();
+                })
+                .setNegativeButton("cancel", new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int which) {
+
+                        dialog.dismiss();
+
+                    }
+                })
+                .create();
+
+        myQuittingDialogBox.show();
     }
 
     private boolean deleteMedsFromAll()
