+++ app/src/main/java/my/bop/finalassignment/ClientFinaliseFeedbackPage.java
@@ -1,7 +1,9 @@
 package my.bop.finalassignment;
 
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
@@ -474,11 +476,31 @@ public class ClientFinaliseFeedbackPage extends AppCompatActivity
 
     public void deleteFeedBack(View view)
     {
-        deletefeedback();
-        Intent intent = new Intent(getApplicationContext(), ClientFinaliseFeedbackPage.class);
-        intent.putExtra("mrn", mrn);
-        intent.putExtra("admissionid", admissionid);
-        startActivity(intent);
+
+        AlertDialog myQuittingDialogBox = new AlertDialog.Builder(this)
+                // set message, title, and icon
+                .setTitle("Delete")
+                .setMessage("Do you want to Delete")
+                .setIcon(R.drawable.ic_baseline_delete_forever_24)
+
+                .setPositiveButton("Delete", (dialog, whichButton) -> {
+                    deletefeedback();
+                    Intent intent = new Intent(getApplicationContext(), ClientFinaliseFeedbackPage.class);
+                    intent.putExtra("mrn", mrn);
+                    intent.putExtra("admissionid", admissionid);
+                    startActivity(intent);
+                    dialog.dismiss();
+                })
+                .setNegativeButton("cancel", new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int which) {
+
+                        dialog.dismiss();
+
+                    }
+                })
+                .create();
+
+        myQuittingDialogBox.show();
 
     }
 
@@ -518,18 +540,38 @@ public class ClientFinaliseFeedbackPage extends AppCompatActivity
 
     public void deleteAdmission(View view)
     {
-        if(deleteAdmissionDatabase())
-        {
-            Intent intent = new Intent(getApplicationContext(), ClientHomePage.class);
-            startActivity(intent);
-        }
-        else
-        {
-            Intent intent = new Intent(getApplicationContext(), ClientFinaliseFeedbackPage.class);
-            intent.putExtra("mrn", mrn);
-            intent.putExtra("admissionid", admissionid);
-            startActivity(intent);
-        }
+        AlertDialog myQuittingDialogBox = new AlertDialog.Builder(this)
+                // set message, title, and icon
+                .setTitle("Delete")
+                .setMessage("Do you want to Delete")
+                .setIcon(R.drawable.ic_baseline_delete_forever_24)
+
+                .setPositiveButton("Delete", (dialog, whichButton) -> {
+                    if(deleteAdmissionDatabase())
+                    {
+                        Intent intent = new Intent(getApplicationContext(), ClientHomePage.class);
+                        startActivity(intent);
+                    }
+                    else
+                    {
+                        Intent intent = new Intent(getApplicationContext(), ClientFinaliseFeedbackPage.class);
+                        intent.putExtra("mrn", mrn);
+                        intent.putExtra("admissionid", admissionid);
+                        startActivity(intent);
+                    }
+                    dialog.dismiss();
+                })
+                .setNegativeButton("cancel", new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int which) {
+
+                        dialog.dismiss();
+
+                    }
+                })
+                .create();
+
+        myQuittingDialogBox.show();
+
     }
 
     private boolean deleteAdmissionDatabase()
