+++ app/src/main/java/my/bop/finalassignment/ClientModifyPatientPage.java
@@ -1,7 +1,9 @@
 package my.bop.finalassignment;
 
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
@@ -272,22 +274,43 @@ public class ClientModifyPatientPage extends AppCompatActivity
 
     public void DeletePatientButton(View view)
     {
-        if(deletePatientSingle())
-        {
-            Intent intent = new Intent(getApplicationContext(), ClientSearchPatientPage.class);
-            startActivity(intent);
-        }
-        else
-        {
-            Intent intent = new Intent(getApplicationContext(), ClientModifyPatientPage.class);
-            intent.putExtra("mrn", mrn1);
-            intent.putExtra("age", age1);
-            intent.putExtra("weight", weight1);
-            intent.putExtra("gender", gender1);
-            intent.putExtra("userid", userID);
-            intent.putExtra("patientid", patientID1);
-            startActivity(intent);
-        }
+
+        AlertDialog myQuittingDialogBox = new AlertDialog.Builder(this)
+                // set message, title, and icon
+                .setTitle("Delete")
+                .setMessage("Do you want to Delete")
+                .setIcon(R.drawable.ic_baseline_delete_forever_24)
+
+                .setPositiveButton("Delete", (dialog, whichButton) -> {
+                    if(deletePatientSingle())
+                    {
+                        Intent intent = new Intent(getApplicationContext(), ClientSearchPatientPage.class);
+                        startActivity(intent);
+                    }
+                    else
+                    {
+                        Intent intent = new Intent(getApplicationContext(), ClientModifyPatientPage.class);
+                        intent.putExtra("mrn", mrn1);
+                        intent.putExtra("age", age1);
+                        intent.putExtra("weight", weight1);
+                        intent.putExtra("gender", gender1);
+                        intent.putExtra("userid", userID);
+                        intent.putExtra("patientid", patientID1);
+                        startActivity(intent);
+                    }
+                    dialog.dismiss();
+                })
+                .setNegativeButton("cancel", new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int which) {
+
+                        dialog.dismiss();
+
+                    }
+                })
+                .create();
+
+        myQuittingDialogBox.show();
+
     }
 
     private boolean deletePatientSingle()
@@ -338,25 +361,42 @@ public class ClientModifyPatientPage extends AppCompatActivity
 
     public void DeletePatientAllButton(View view)
     {
-        if(deletePatient())
-        {
-            Intent intent = new Intent(getApplicationContext(), ClientSearchPatientPage.class);
-            startActivity(intent);
-        }
-        else
-        {
-            Intent intent = new Intent(getApplicationContext(), ClientModifyPatientPage.class);
-            intent.putExtra("mrn", mrn1);
-            intent.putExtra("age", age1);
-            intent.putExtra("weight", weight1);
-            intent.putExtra("gender", gender1);
-            intent.putExtra("userid", userID);
-            intent.putExtra("patientid", patientID1);
-            startActivity(intent);
-        }
+        AlertDialog myQuittingDialogBox = new AlertDialog.Builder(this)
+                // set message, title, and icon
+                .setTitle("Delete")
+                .setMessage("Do you want to Delete")
+                .setIcon(R.drawable.ic_baseline_delete_forever_24)
+
+                .setPositiveButton("Delete", (dialog, whichButton) -> {
+                    if(deletePatient())
+                    {
+                        Intent intent = new Intent(getApplicationContext(), ClientSearchPatientPage.class);
+                        startActivity(intent);
+                    }
+                    else
+                    {
+                        Intent intent = new Intent(getApplicationContext(), ClientModifyPatientPage.class);
+                        intent.putExtra("mrn", mrn1);
+                        intent.putExtra("age", age1);
+                        intent.putExtra("weight", weight1);
+                        intent.putExtra("gender", gender1);
+                        intent.putExtra("userid", userID);
+                        intent.putExtra("patientid", patientID1);
+                        startActivity(intent);
+                    }
+                    dialog.dismiss();
+                })
+                .setNegativeButton("cancel", new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int which) {
+
+                        dialog.dismiss();
+
+                    }
+                })
+                .create();
+
+        myQuittingDialogBox.show();
     }
-
-
 
     public void onClickNavigation(MenuItem item)
     {
